requirements.txt
Django==5.0
mysqlclient==2.2.0
Pillow==10.2.0
python-decouple==3.8
google-generativeai==0.8.3
qrcode[pil]==7.4.2
